#56 chore: 구독 API url 계층 구조 변경
