+++ src/main/java/com/ewhatever/qna/mail/service/MailService.java
@@ -1,13 +1,7 @@
 package com.ewhatever.qna.mail.service;
 
 import com.ewhatever.qna.common.Base.BaseException;
-import com.ewhatever.qna.login.CustomUnauthorizedException;
-import com.ewhatever.qna.login.JwtIssuer;
-import com.ewhatever.qna.login.dto.AuthService;
-import com.ewhatever.qna.mail.dto.GetSubscriptionRes;
 import com.ewhatever.qna.mail.dto.PostMailReq;
-import com.ewhatever.qna.mail.dto.PostSubscriptionReq;
-import com.ewhatever.qna.user.entity.User;
 import com.ewhatever.qna.user.repository.UserRepository;
 import jakarta.mail.MessagingException;
 import jakarta.mail.internet.MimeMessage;
@@ -17,8 +11,6 @@ import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.mail.javamail.MimeMessageHelper;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.validation.BindingResult;
 import org.thymeleaf.context.Context;
 import org.thymeleaf.spring6.SpringTemplateEngine;
 
@@ -35,13 +27,7 @@ public class MailService {
 
     private JavaMailSender emailSender;
     private final SpringTemplateEngine templateEngine;
-
-    private final JwtIssuer jwtIssuer;
-
     private final UserRepository userRepository;
-
-    private final AuthService authService;
-
     @Async
     public void sendMail(PostMailReq postMailReq) throws BaseException {
 
@@ -80,24 +66,5 @@ public class MailService {
         }
 
             log.info("*** 메일 전송 완료 ***");
-    }
-
-    @Transactional
-    public void subscribeLetter(String token, PostSubscriptionReq postSubscriptionReq, BindingResult bindingResult) throws BaseException {
-        if(!jwtIssuer.validateToken(token)) throw new CustomUnauthorizedException(INVALID_TOKEN.getMessage());
-        if(bindingResult.hasErrors()) throw new BaseException(INVALID_EMAIL);
-        User user = userRepository.findById(authService.getUserIdx(token)).orElseThrow(()-> new BaseException(INVALID_USER));
-        user.setEmail(postSubscriptionReq.getEmail());
-    }
-
-    @Transactional
-    public void deleteSubscriptionLetter(String token) throws BaseException {
-        User user = userRepository.findById(authService.getUserIdx(token)).orElseThrow(()-> new BaseException(INVALID_USER));
-        user.setEmail(null);
-    }
-
-    public GetSubscriptionRes getSubscriptionInformation(String token) throws BaseException {
-        User user = userRepository.findById(authService.getUserIdx(token)).orElseThrow(()-> new BaseException(INVALID_USER));
-        return new GetSubscriptionRes(user);
     }
 }
